healthcare-translator/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
groq==0.15.0
python-multipart==0.0.20
python-dotenv==1.0.1
websockets==14.1
aiofiles==24.1.0
psycopg2-binary==2.9.10
edge-tts==6.1.18
gtts 

